requirements.txt
Pillow
pyperclip